requirements.txt
python-telegram-bot>=20.0
ccxt>=4.0.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.28.0
python-dotenv>=1.0.0